requirements.txt
Pyrogram==2.0.30
python-dotenv==0.20.0
TgCrypto==1.2.3